--- Source/PhoneBook/src/main/java/org/example/Controller/DatabaseExtention.java
@@ -1,29 +1,0 @@
-package org.example.Controller;
-
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public abstract class DatabaseExtention {
-    private Connection con; //connection
-    public DatabaseExtention(Connection con){
-        this.con = con;
-    }
-    public Connection getCon(){
-        return con;
-    } //getting connection but rn idk if this is neccesary lmao
-    public void printResult(ResultSet rs) throws SQLException { //printing given result
-        int rowCount = 0;
-        rs.last();
-        rowCount = rs.getRow();
-        rs.beforeFirst();
-
-        while(rs.next()){
-            String out = "";
-            for(int x = 1; x < rowCount; x++){
-                out += rs.getString(x) + " ";
-            }
-            System.out.println(out);
-        }
-    }
-}
